nft721 sales flow working
