+++ src/keeper/contracts/templates/BaseTemplate.abstract.ts
@@ -1,7 +1,9 @@
 import { AgreementTemplate } from './AgreementTemplate.abstract'
-import { zeroX } from '../../../utils'
+import { ZeroAddress, zeroX } from '../../../utils'
 import Account from '../../../nevermined/Account'
 import { TxParameters } from '../ContractBase'
+import Token from '../Token'
+import CustomToken from '../CustomToken'
 
 export abstract class BaseTemplate extends AgreementTemplate {
     /**
@@ -76,5 +78,41 @@ export abstract class BaseTemplate extends AgreementTemplate {
         agreementId: string
     ): Promise<{ accessProvider: string; accessConsumer: string }> {
         return this.call<any>('getAgreementData', [zeroX(agreementId)])
+    }
+
+    public async lockTokens(tokenAddress, amounts, from: Account, txParams: TxParameters): Promise<void> {
+        let token: Token
+
+        const {
+            lockPaymentCondition,
+        } = this.nevermined.keeper.conditions
+
+        if (!tokenAddress) {
+            ;({ token } = this.nevermined.keeper)
+        } else if (tokenAddress.toLowerCase() !== ZeroAddress) {
+            token = await CustomToken.getInstanceByAddress(
+                {
+                    nevermined: this.nevermined,
+                    web3: this.web3,
+                    logger: this.logger,
+                    config: this.config
+                },
+                tokenAddress
+            )
+        }
+
+        const totalAmount = amounts.reduce((a, b) => a + b, 0)
+
+        if (token) {
+            this.logger.debug('Approving tokens', totalAmount)
+            await token.approve(
+                lockPaymentCondition.getAddress(),
+                totalAmount,
+                from,
+                txParams
+            )
+        }
+
+
     }
 }
