+++ src/keeper/contracts/templates/NFT721SalesTemplate.ts
@@ -71,37 +71,7 @@ export class NFT721SalesTemplate extends BaseTemplate {
             consumerAddress.getId()
         )
 
-        let token: Token
-
-        const {
-            lockPaymentCondition,
-        } = this.nevermined.keeper.conditions
-
-        if (!tokenAddress) {
-            ;({ token } = this.nevermined.keeper)
-        } else if (tokenAddress.toLowerCase() !== ZeroAddress) {
-            token = await CustomToken.getInstanceByAddress(
-                {
-                    nevermined: this.nevermined,
-                    web3: this.web3,
-                    logger: this.logger,
-                    config: this.config
-                },
-                tokenAddress
-            )
-        }
-
-        const totalAmount = amounts.reduce((a, b) => a + b, 0)
-
-        if (token) {
-            this.logger.debug('Approving tokens', totalAmount)
-            await token.approve(
-                lockPaymentCondition.getAddress(),
-                totalAmount,
-                from,
-                txParams
-            )
-        }
+        await this.lockTokens(tokenAddress, amounts, from, txParams)
 
         return !!(await this.createAgreementAndPay(
             agreementId,
